litebastion: Support requests on a local endpoint
This commit makes it possible to receive backend connections from the
public internet while *only* accepting requests to those backends on a
different endpoint (forced on 127.0.0.1 because of being HTTP only).

This is for example useful when running a bastion host on the same
system as a log; or when the bastion host operator has its own
authenticated interface to inject the -local-request-at endpoint.